[refactor] Header components로 파일 이전
