+++ apps/frontend/src/components/layout/AppLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 
-import Header from './Header';
+import Header from '../ui/Header';
 import Sidebar from '../ui/Sidebar';
 
 export default function AppLayout() {
